LMS-20:updated user model and user dto (#27)

+++ src/main/java/com/java/librarymanagement/users/model/UserDTO.java
@@ -11,7 +11,7 @@ public class UserDTO {
     private long id;
     private String name;
     private String email;
-    private Long phoneNumber;
+    private Integer phoneNumber;
     private String course;
     private String roles;
 }
